Move the commands in the base command package to better locations
