+++ src/main/java/bhs/devilbotz/RobotContainer.java
@@ -5,10 +5,10 @@
 
 package bhs.devilbotz;
 
-import bhs.devilbotz.commands.BalancePID;
+import bhs.devilbotz.commands.auto.BalancePID;
 import bhs.devilbotz.commands.DriveCommand;
-import bhs.devilbotz.commands.DriveStraightPID;
-import bhs.devilbotz.commands.DriveStraightToDock;
+import bhs.devilbotz.commands.auto.DriveStraightPID;
+import bhs.devilbotz.commands.auto.DriveStraightToDock;
 import bhs.devilbotz.commands.arm.ArmDown;
 import bhs.devilbotz.commands.arm.ArmMoveDistance;
 import bhs.devilbotz.commands.arm.ArmStop;
