+++ src/main/java/bhs/devilbotz/commands/auto/DriveStraightPID.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package bhs.devilbotz.commands;
+package bhs.devilbotz.commands.auto;
 
 import bhs.devilbotz.Robot;
 import bhs.devilbotz.subsystems.DriveTrain;
